refactor: use store
